[MO-154] Move all react-select custom components to their own files
