+++ src/ui/Forms/Select/ClearIndicator.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const StyledClearIndicator = styled.div``;
+
+const ClearIndicator = props => {
+  return <StyledClearIndicator>Cancel</StyledClearIndicator>;
+};
+
+export default ClearIndicator;
