+++ src/ui/Forms/Select/LoadingIndicator.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Loader from 'react-loader-spinner';
+
+const LoadingIndicator = () => {
+  return <Loader type="TailSpin" color="#00BFFF" height="13" width="13" />;
+};
+
+export default LoadingIndicator;
