+++ src/ui/Forms/Select/Value.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+import { rem } from 'polished';
+
+const Value = styled.div`
+  display: flex;
+  align-content: center;
+  flex-wrap: wrap;
+  padding: ${rem('8px')} 0;
+  flex: 1;
+`;
+
+export default Value;
